[Decomp][Pet] Cohesion split: extract persistence + spells/talents to Pet*.cpp (#265)
Re-applies the mangosthree decomposition onto MangosOne's own TBC 2.4.3
source. Pure code move out of the parent translation unit into the new
cohesively-named file(s); no behaviour change (source diff is deletion-only).
CMake file(GLOB) picks the new TU(s) up automatically. Verified: game.lib
compiles and links clean (MSVC, Release).

Not ported (absent in MangosOne TBC 2.4.3 -- these are
later-expansion methods that do not exist in this branch):

Pet::InitLevelupSpellsForLevel,Pet::resetTalents,Pet::resetTalentsForAllPetsOf,Pet::UpdateFreeTalentPoints,Pet::InitTalentForLevel,Pet::GetMaxTalentPointsForLevel
Pet::InitLevelupSpellsForLevel,Pet::resetTalents,Pet::resetTalentsForAllPetsOf,Pet::UpdateFreeTalentPoints,Pet::InitTalentForLevel,Pet::GetMaxTalentPointsForLevel
Ported-from: mangosthree/server@1bae7aac0651
